Previous project(s)
This project was an RIS system from software engineering. We had some framework setup which involved storing data from a user interface, passing it through java API and into MySQL. Could be a helpful reference for us.
The labTwo zip is all the way from 1302, so I am not sure how useful it will be. But, it was an inventory organizing program for walmart. It is very poor in quality from what i remember, but something is better than nothing.